'grunt cordova' command is not clear
Most commands have their own line and are clear when someone is skimming through the documentation, however 'grunt cordova' is easily missed as it is inline with the instructions. It would be more consistent to have it on it's own line so that it's easily detectable. I missed it and it took me a few times to realize it.