Update install_plugins.js
I've got some issues when the plugins are installed asynchronously, (some plugins are installed in the iOS project, some not), this is fixed if the plugin install process is made synchronously. This is slower, but fixes the problem, and I think its not a huge issue to wait some extra seconds.